Cleanup files into a package like format
